@@ -1,19 +1,22 @@
 
 package com.drazail.rnandroidstore;
 
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+import java.util.ArrayList;
 
 import com.facebook.react.ReactPackage;
 import com.facebook.react.bridge.NativeModule;
 import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.uimanager.ViewManager;
 import com.facebook.react.bridge.JavaScriptModule;
+
 public class RNAndroidStorePackage implements ReactPackage {
     @Override
     public List<NativeModule> createNativeModules(ReactApplicationContext reactContext) {
-      return Arrays.<NativeModule>asList(new RNAndroidStore(reactContext));
+      List<NativeModule> modules = new ArrayList<>();
+      modules.add(new RNAndroidStore(reactContext));
+      return modules;
     }
 
     // Deprecated from RN 0.47
